Extract moment from util
[ci skip]
