+++ lib/ulid.rb
@@ -8,6 +8,7 @@ require('securerandom')
 require_relative('ulid/version')
 require_relative('ulid/errors')
 require_relative('ulid/crockford_base32')
+require_relative('ulid/moment')
 require_relative('ulid/utils')
 require_relative('ulid/uuid')
 require_relative('ulid/monotonic_generator')
